+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 0.2.0 *(2015-03-xx)*
 
  * Refactor to support YTS Rest update to v2. This release contains **breaking changes**.
  * The endpoints have been divided into different modules based on the functionality.
- * All methods are now overloaded to return an rx.Observable.
+ * All methods are now overloaded to return an `rx.Observable`.
  * Update to Gradle wrapper 2.2.1
  * Update to Retrofit 1.9.0